# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  employeeId?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  jobPosition?: string | null;
  /** @nullable */
  company?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface LoginInput {
  loginId: string;
  password: string;
}

export interface SignupInput {
  name: string;
  email: string;
  password: string;
  role: string;
  companyName: string;
  department: string;
  jobPosition: string;
  phone: string;
  address: string;
}

export interface Employee {
  id: string;
  name: string;
  email: string;
  employeeId: string;
  department: string;
  jobPosition: string;
  status: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  joiningDate?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  avatar?: string | null;
}

export interface EmployeeUpdate {
  name?: string;
  department?: string;
  jobPosition?: string;
  phone?: string;
  address?: string;
}

export interface Attendance {
  id: string;
  date: string;
  status: string;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  workHours?: number | null;
  /** @nullable */
  userId?: string | null;
}

export interface Leave {
  id: string;
  type: string;
  startDate: string;
  endDate: string;
  status: string;
  reason: string;
  /** @nullable */
  adminComment?: string | null;
  user?: Employee;
}

export interface LeaveInput {
  type: string;
  startDate: string;
  endDate: string;
  reason: string;
  attachmentUrl?: string;
}

export interface LeaveStatusInput {
  status: string;
  adminComment?: string;
}

export interface LeaveBalance {
  type: string;
  total: number;
  used: number;
  available: number;
}

export type PayrollEarnings = {[key: string]: number};

export type PayrollDeductions = {[key: string]: number};

export interface Payroll {
  monthlyWage: number;
  yearlyWage: number;
  netSalary: number;
  earnings: PayrollEarnings;
  deductions: PayrollDeductions;
}

export interface PayrollUpdate {
  monthlyWage: number;
}

export type DashboardStatsDepartmentDistributionItem = {
  department: string;
  count: number;
};

export type DashboardStatsAttendanceTrendItem = {
  date: string;
  present: number;
  absent: number;
};

export interface DashboardStats {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  onLeave: number;
  pendingLeaves: number;
  departmentDistribution: DashboardStatsDepartmentDistributionItem[];
  attendanceTrend: DashboardStatsAttendanceTrendItem[];
}

export type GetEmployeesParams = {
search?: string;
status?: string;
department?: string;
};

export type GetAttendanceLogsParams = {
startDate?: string;
endDate?: string;
userId?: string;
};