# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dayflow HRMS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: dashboard
  - name: employees
  - name: attendance
  - name: leaves
  - name: payroll
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: department
          schema: { type: string }
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Employee" }
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
    put:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeUpdate" }
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      responses:
        "200":
          description: Attendance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      responses:
        "200":
          description: Attendance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
  /attendance/today:
    get:
      operationId: getAttendanceToday
      tags: [attendance]
      responses:
        "200":
          description: Today's attendance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
  /attendance/logs:
    get:
      operationId: getAttendanceLogs
      tags: [attendance]
      parameters:
        - in: query
          name: startDate
          schema: { type: string }
        - in: query
          name: endDate
          schema: { type: string }
        - in: query
          name: userId
          schema: { type: string }
      responses:
        "200":
          description: Attendance logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Attendance" }
  /leaves:
    get:
      operationId: getLeaves
      tags: [leaves]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Leave" }
  /leaves/apply:
    post:
      operationId: applyLeave
      tags: [leaves]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveInput" }
      responses:
        "201":
          description: Leave request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Leave" }
  /leaves/balance:
    get:
      operationId: getLeaveBalance
      tags: [leaves]
      responses:
        "200":
          description: Leave balances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaveBalance" }
  /leaves/{id}/status:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveStatusInput" }
      responses:
        "200":
          description: Updated leave request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Leave" }
  /payroll/my-slip:
    get:
      operationId: getMyPayroll
      tags: [payroll]
      responses:
        "200":
          description: Payroll slip
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payroll" }
  /payroll/{userId}:
    put:
      operationId: updatePayroll
      tags: [payroll]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayrollUpdate" }
      responses:
        "200":
          description: Updated payroll
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payroll" }
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        employeeId: { type: [string, "null"] }
        avatar: { type: [string, "null"] }
        department: { type: [string, "null"] }
        jobPosition: { type: [string, "null"] }
        company: { type: [string, "null"] }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    LoginInput:
      type: object
      required: [loginId, password]
      properties:
        loginId: { type: string }
        password: { type: string }
    SignupInput:
      type: object
      required: [name, email, password, role, companyName, department, jobPosition, phone, address]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { type: string }
        companyName: { type: string }
        department: { type: string }
        jobPosition: { type: string }
        phone: { type: string }
        address: { type: string }
    Employee:
      type: object
      required: [id, name, email, employeeId, department, jobPosition, status]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        employeeId: { type: string }
        department: { type: string }
        jobPosition: { type: string }
        status: { type: string }
        phone: { type: [string, "null"] }
        address: { type: [string, "null"] }
        joiningDate: { type: [string, "null"] }
        role: { type: [string, "null"] }
        avatar: { type: [string, "null"] }
    EmployeeUpdate:
      type: object
      properties:
        name: { type: string }
        department: { type: string }
        jobPosition: { type: string }
        phone: { type: string }
        address: { type: string }
    Attendance:
      type: object
      required: [id, date, status]
      properties:
        id: { type: string }
        date: { type: string }
        status: { type: string }
        checkIn: { type: [string, "null"] }
        checkOut: { type: [string, "null"] }
        workHours: { type: [number, "null"] }
        userId: { type: [string, "null"] }
    Leave:
      type: object
      required: [id, type, startDate, endDate, status, reason]
      properties:
        id: { type: string }
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        status: { type: string }
        reason: { type: string }
        adminComment: { type: [string, "null"] }
        user: { $ref: "#/components/schemas/Employee" }
    LeaveInput:
      type: object
      required: [type, startDate, endDate, reason]
      properties:
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        reason: { type: string }
        attachmentUrl: { type: string }
    LeaveStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string }
        adminComment: { type: string }
    LeaveBalance:
      type: object
      required: [type, total, used, available]
      properties:
        type: { type: string }
        total: { type: number }
        used: { type: number }
        available: { type: number }
    Payroll:
      type: object
      required: [monthlyWage, yearlyWage, netSalary, earnings, deductions]
      properties:
        monthlyWage: { type: number }
        yearlyWage: { type: number }
        netSalary: { type: number }
        earnings:
          type: object
          additionalProperties: { type: number }
        deductions:
          type: object
          additionalProperties: { type: number }
    PayrollUpdate:
      type: object
      required: [monthlyWage]
      properties:
        monthlyWage: { type: number }
    DashboardStats:
      type: object
      required: [totalEmployees, presentToday, absentToday, onLeave, pendingLeaves, departmentDistribution, attendanceTrend]
      properties:
        totalEmployees: { type: number }
        presentToday: { type: number }
        absentToday: { type: number }
        onLeave: { type: number }
        pendingLeaves: { type: number }
        departmentDistribution:
          type: array
          items:
            type: object
            required: [department, count]
            properties:
              department: { type: string }
              count: { type: number }
        attendanceTrend:
          type: array
          items:
            type: object
            required: [date, present, absent]
            properties:
              date: { type: string }
              present: { type: number }
              absent: { type: number }