# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendance {
  id: string;
  date: string;
  status: string;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  workHours?: number | null;
  /** @nullable */
  userId?: string | null;
}