# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsAttendanceTrendItem } from './dashboardStatsAttendanceTrendItem';
import type { DashboardStatsDepartmentDistributionItem } from './dashboardStatsDepartmentDistributionItem';

export interface DashboardStats {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  onLeave: number;
  pendingLeaves: number;
  departmentDistribution: DashboardStatsDepartmentDistributionItem[];
  attendanceTrend: DashboardStatsAttendanceTrendItem[];
}