# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: string;
  name: string;
  email: string;
  employeeId: string;
  department: string;
  jobPosition: string;
  status: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  joiningDate?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  avatar?: string | null;
}