# lib/api-zod/src/generated/types/employeeUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeUpdate {
  name?: string;
  department?: string;
  jobPosition?: string;
  phone?: string;
  address?: string;
}