# lib/api-zod/src/generated/types/leave.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';

export interface Leave {
  id: string;
  type: string;
  startDate: string;
  endDate: string;
  status: string;
  reason: string;
  /** @nullable */
  adminComment?: string | null;
  user?: Employee;
}