# lib/api-zod/src/generated/types/leaveBalance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveBalance {
  type: string;
  total: number;
  used: number;
  available: number;
}