# lib/api-zod/src/generated/types/leaveInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveInput {
  type: string;
  startDate: string;
  endDate: string;
  reason: string;
  attachmentUrl?: string;
}