# lib/api-zod/src/generated/types/leaveStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveStatusInput {
  status: string;
  adminComment?: string;
}