# lib/api-zod/src/generated/types/payroll.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */
import type { PayrollDeductions } from './payrollDeductions';
import type { PayrollEarnings } from './payrollEarnings';

export interface Payroll {
  monthlyWage: number;
  yearlyWage: number;
  netSalary: number;
  earnings: PayrollEarnings;
  deductions: PayrollDeductions;
}