# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  name: string;
  email: string;
  password: string;
  role: string;
  companyName: string;
  department: string;
  jobPosition: string;
  phone: string;
  address: string;
}