# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dayflow HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  employeeId?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  jobPosition?: string | null;
  /** @nullable */
  company?: string | null;
}